requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
geoip2==5.0.1
idna==3.10
maxminddb==2.6.3
multidict==6.2.0
packaging==24.2
pefile==2023.2.7
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
setuptools==77.0.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
